// vietnam/assets/ScriptCore/protocol/shop.d.ts
export namespace shop {
// ...
    interface IShopInfo {

        /** ShopInfo Id */
        Id?: (number|null);

        /** ShopInfo AdLookedNum */
        AdLookedNum?: (number|null);

        /** ShopInfo AdReceiveNum */
        AdReceiveNum?: (number|null);

        /** ShopInfo LastLookTime */
        LastLookTime?: (number|null);

        /** ShopInfo RoleAdded */
        RoleAdded?: (number|null);

        /** ShopInfo PetAdded */
        PetAdded?: (number|null);
    }
// ...
    interface IShopExchangeInfo {

        /** ShopExchangeInfo Type */
        Type?: (number|null);
// ...
    }
// ...
}